# api/swagger/gasbank_endpoints.yaml
openapi: 3.0.0
info:
  title: Service Layer Gas Bank API
  description: API documentation for the Gas Bank service endpoints
  version: 1.0.0
  contact:
    name: R3E Network
    email: support@r3e-network.io

servers:
  - url: https://api.service-layer.io/v1
    description: Production server
  - url: https://staging-api.service-layer.io/v1
    description: Staging server
  - url: http://localhost:8080/v1
    description: Local development server

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key

  schemas:
    GasAccountBalance:
      type: object
      properties:
        account_address:
          type: string
          description: NEO account address
          example: "NXwQ5UWPJYjkV3MYs3cQNrzdgwpbTkMELN"
        gas_balance:
          type: number
          format: float
          description: Account GAS balance
          example: 42.5
        reserved_gas:
          type: number
          format: float
          description: Reserved GAS for pending operations
          example: 10.0
        available_gas:
          type: number
          format: float
          description: Available GAS for new operations
          example: 32.5
        last_updated:
          type: string
          format: date-time
          description: Timestamp of last balance update
          example: "2023-08-15T12:34:56Z"

    GasTransferRequest:
      type: object
      required:
        - destination_address
        - amount
      properties:
        destination_address:
          type: string
          description: Destination NEO address
          example: "NXwQ5UWPJYjkV3MYs3cQNrzdgwpbTkMELN"
        amount:
          type: number
          format: float
          description: Amount of GAS to transfer
          example: 5.0
        memo:
          type: string
          description: Optional transaction memo
          example: "Payment for service"

    GasTransaction:
      type: object
      properties:
        id:
          type: string
          description: Transaction ID
          example: "tx-123456"
        user_id:
          type: integer
          description: User ID
          example: 42
        tx_hash:
          type: string
          description: Blockchain transaction hash
          example: "0x1234567890abcdef1234567890abcdef1234567890abcdef1234567890abcdef"
        type:
          type: string
          enum: [deposit, withdrawal, service_fee, refund]
          description: Transaction type
          example: "deposit"
        amount:
          type: number
          format: float
          description: GAS amount
          example: 5.0
        status:
          type: string
          enum: [pending, confirmed, failed]
          description: Transaction status
          example: "confirmed"
        created_at:
          type: string
          format: date-time
          description: Transaction creation timestamp
          example: "2023-08-15T12:34:56Z"
        updated_at:
          type: string
          format: date-time
          description: Transaction last update timestamp
          example: "2023-08-15T12:40:22Z"
        memo:
          type: string
          description: Transaction memo
          example: "Deposit from personal wallet"

paths:
  /gasbank/balance:
    get:
      summary: Get gas balance
      description: Returns the current gas balance for the authenticated user
      operationId: getGasBalance
      tags:
        - Gas Bank
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      responses:
        '200':
          description: Gas balance information
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/GasAccountBalance'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '../openapi.yaml#/components/schemas/Error'

  /gasbank/deposit-address:
    get:
      summary: Get deposit address
      description: Returns a deposit address for adding GAS to the user's account
      operationId: getDepositAddress
      tags:
        - Gas Bank
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      responses:
        '200':
          description: Deposit address information
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      deposit_address:
                        type: string
                        description: NEO address for depositing GAS
                        example: "NXwQ5UWPJYjkV3MYs3cQNrzdgwpbTkMELN"
                      qr_code_url:
                        type: string
                        description: URL to QR code image
                        example: "https://api.service-layer.io/v1/gasbank/qr/NXwQ5UWPJYjkV3MYs3cQNrzdgwpbTkMELN"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '../openapi.yaml#/components/schemas/Error'

  /gasbank/withdraw:
    post:
      summary: Withdraw GAS
      description: Withdraws GAS from the user's account to an external address
      operationId: withdrawGas
      tags:
        - Gas Bank
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GasTransferRequest'
      responses:
        '200':
          description: Withdrawal request successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      transaction_id:
                        type: string
                        description: Withdrawal transaction ID
                        example: "tx-123456"
                      status:
                        type: string
                        enum: [pending, confirmed, failed]
                        description: Transaction status
                        example: "pending"
                      amount:
                        type: number
                        format: float
                        description: Withdrawal amount
                        example: 5.0
                      fee:
                        type: number
                        format: float
                        description: Withdrawal fee
                        example: 0.1
                      destination_address:
                        type: string
                        description: Destination address
                        example: "NXwQ5UWPJYjkV3MYs3cQNrzdgwpbTkMELN"
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '../openapi.yaml#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '../openapi.yaml#/components/schemas/Error'
        '402':
          description: Insufficient funds
          content:
            application/json:
              schema:
                $ref: '../openapi.yaml#/components/schemas/Error'

  /gasbank/transactions:
    get:
      summary: Get gas transactions
      description: Returns a list of gas transactions for the authenticated user
      operationId: getGasTransactions
      tags:
        - Gas Bank
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      parameters:
        - name: page
          in: query
          description: Page number for pagination
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: limit
          in: query
          description: Number of items per page
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - name: type
          in: query
          description: Filter by transaction type
          schema:
            type: string
            enum: [deposit, withdrawal, service_fee, refund]
        - name: status
          in: query
          description: Filter by transaction status
          schema:
            type: string
            enum: [pending, confirmed, failed]
      responses:
        '200':
          description: List of gas transactions
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      transactions:
                        type: array
                        items:
                          $ref: '#/components/schemas/GasTransaction'
                      pagination:
                        type: object
                        properties:
                          total:
                            type: integer
                            description: Total number of transactions
                            example: 42
                          page:
                            type: integer
                            description: Current page number
                            example: 1
                          limit:
                            type: integer
                            description: Number of items per page
                            example: 20
                          pages:
                            type: integer
                            description: Total number of pages
                            example: 3
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '../openapi.yaml#/components/schemas/Error'

  /gasbank/transactions/{id}:
    get:
      summary: Get transaction details
      description: Returns details for a specific gas transaction
      operationId: getTransactionDetails
      tags:
        - Gas Bank
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      parameters:
        - name: id
          in: path
          description: Transaction ID
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Transaction details
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/GasTransaction'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '../openapi.yaml#/components/schemas/Error'
        '404':
          description: Transaction not found
          content:
            application/json:
              schema:
                $ref: '../openapi.yaml#/components/schemas/Error' 